+++ src/preload/index.d.ts
@@ -21,6 +21,7 @@ import type {
   VersionsInfo,
   EditFormat,
   OS,
+  ModelInfo,
 } from '@common/types';
 
 export interface ApplicationAPI {
@@ -79,6 +80,7 @@ export interface ApplicationAPI {
   getReleaseNotes: () => Promise<string | null>;
   clearReleaseNotes: () => Promise<void>;
   getOS: () => Promise<OS>;
+  loadModelsInfo: () => Promise<Record<string, ModelInfo>>;
 
   addResponseChunkListener: (baseDir: string, callback: (event: Electron.IpcRendererEvent, data: ResponseChunkData) => void) => string;
   removeResponseChunkListener: (listenerId: string) => void;