+++ src/preload/index.ts
@@ -98,6 +98,7 @@ const api: ApplicationAPI = {
   getReleaseNotes: () => ipcRenderer.invoke('get-release-notes'),
   clearReleaseNotes: () => ipcRenderer.invoke('clear-release-notes'),
   getOS: (): Promise<OS> => ipcRenderer.invoke('get-os'),
+  loadModelsInfo: () => ipcRenderer.invoke('load-models-info'),
 
   addResponseChunkListener: (baseDir, callback) => {
     const listenerId = uuidv4();